Refactor to create groups instead of maps/arrays
This is useful to have as it lets us have one representation for both
map and array represented groups, which can differ by just
serializaiton in the future.

This also allows for easier creation of named nested inline groups
which will greatly simplify the implementation of codegen for nested
groups.

The future wrappers over this group representation will be defined
at the root level and have wasm exposure, rather than being hidden in
the group module.