Inline 1-element group choices
This is instead of generating unneccessary group choice variants that
simply wrap the single field.

The variant will also use the field name (if provided), or else type
name as the variant name, allowing a way to have control over variant
names by making all group choices a separate group and then putting them
as an identifier in the choice.