Basic null support
Extends RustType to take on fixed-value ie a fixed u32, text, or null.
This allows us to have null as type choices or as field types while
refactoring the existing code for fixed uint values to make it easier to
extend.

The only issue we have is that now `RustType::for_member()` can fail if
it is `RustType::Fixed``. This should probably be refactored to return
an error type or something, or split RustType off into something like:
```
enum FieldType {
    RustType(RustType),
    Fixed(FixedValue),
}
```