int type (int = uint / nint)
needed for tx metdata

also moved check for plain groups not having choices to only when we're
trying to generate new functions for them which happens when they're in
an enum or inlined into another group choice.

Before there was a problem with `address` as it had multiple choices but
was never used here so it made no sense to have the check for all plain
groups.