Deserialize nested plain groups inside arrays properly
We needed to read the break for the indefinite case always in the nested
one even when called from deserialize_embedded_group  or else it will be
read from within the outer group.

This caused issues when you had a plain group inside an array group
where it wasn't the last element.

e.g.
```
foo = (uint, text)
bar = [1, foo, 2]
```
it would then read the Special Break from foo instead of the 2 in the
case where foo was encoded with indefinite length encoding.