Misc fixes + multi-layer pass for no-deserialize generation
Out of order declarations like:
```
foo = [bar]
bar = [uint, uint]
```

when bar did not support deserialization were broken before since
we processed in order.

It's still done in order, however we do a super basic multi-pass over
all definitions.

It's not the best, but it works enough as-is as long as there's no
obscenely long chains that are out of order.