Fix deserialization for map-represented structs
Map structs read the ending Special::Break to know that they are ending,
but we were not checking what kind of Special we encountered since we
don't support Special keys in maps (only true/false make sense anyway)
but that meant that we didn't consume the Special CBOR tag, which caused
problems if a map-represented struct was embedded within another CDDL
structure, as the next structure would read the Special::Break instead
of its own actual first value.

Now we consume the Special value we encounter and throw an error if it
is not Special::Break.