Switch indefinite -> definite encodings + only generate SerializeEmbeddedGroup impl when necessary
Shelley Haskell node doesn't support indefinite length deserialization
everywhere for CBOR structures, so to avoid possible problems down the
road, we are switching to definite length encodings. This also should
result in a slight (~5% for txs approx) size savings as indefinite is 2
bytes always but definite is 1 byte for <24 and 2 bytes for I think <256
lengths.

As the serialization impl had to be refactored to do this, I also
decided to only generated the SerializeEmbeddedGroup impl when necessary
(plain groups only) and just directly doing everything in Serialize
instead of delegating off to SerializeEmbeddedGroup within every single
Serilize. This should make things more clear if anyone ever tries to use
the ruse seriallization code directly so they don't mess up and not
include tags/array tags/etc. It also results in smaller code with less
size bloat.

In the future it should be easy as well to have some config that could
decide between definite vs indefinite as a result of this refactor.