Change wasm-exposed interface to be friendlier
Now all wasm-exposed functions take by-reference rust-structs and clone
them so that they can't be moved in rust memory which is not obvious
from the js/wasm side. Now everything should be ref+clone()

The BINARY_WRAPPER flag was also added to toggle the behavior of
generating binary wrappers for bytes/bstr aliases.

RustType also now has an Alias variant for when it was the produce of a
user-specified type alias. For reserved keywords this is not used.