Generate getters
All non-fixed values generate a getter

All enums (group choices + type choices) now generate a FooKind in
addition to the FooEnum which  is wasm-exposed and contains no data, just
specifies which FooEnum the Foo struct is.

These enums also, for data-containing variants, generate an as_* variant
conversion function which returns a copy of the variant if it is that
variant, or else None (undefined from js/wasm) if it's another variant.

A config arg for GENERATE_TO_FROM_BYTES was also added to toggle the
generation of to_bytes() and from_bytes().

Deserialize was added to Int as well.