Auto-generate cddl prelude definitions when needed
The CDDL spec defines many types in its standard prelude which we don't
have a hand-tailored solution for, unlike with bool, int, etc.
In these cases we use the cddl definition from the cddl RFC to parse
them upon being needed giving us very direct types for them.

One difference is that we prefix them all with prelude_ to get around a
restriction in the cddl parsing lib we are using which errors out if any
prelude identifier is the left hand side of a type rule.

We will likely want to continue hand-implementing some to make them
easier to use instead of heavily relying on type choices and many
definitions for each type. The big int types could likely be combined
and given useful functions.