Respect deserialized format (e.g. order, en encodings)
CBOR has several ways to vary in the bytes encoded from a single
semantically-equal piece of information:

1) Map ordering of keys
2) Integer encodings (1 vs 4 bytes) (includes length for major types)
3) Definite vs indefinite lengths of arrays/maps (and strings)

We implement 1 and 3 (excluding for strings) here as 2 would require
changes to cbor_event and for our purposes (Cardano blockchain) we've
never seen any data being posted that didn't use the smallest encoding
for integers/lengths.

This is controlled by a new flag in `cmd.rs`. If this is not something
an end-user cares about they should disable it as it creates extra
fields to remember this data as well as extra code. It also necessitates
an extra dependency for `linked_hash_map`.