Rust / Wasm Split Initial Work
splits export into two crates: core/ and wasm/

Covers everything except entirely splitting out Vec<T> / Map<K, V>
into pure rust to avoid having wrapper types. The table/array code is
semi in-flux as of this commit, but will be fully resolved in a later
commit.

The previous tests (besides rust_wasm_split) need to be updated still,
as well as tests for the wasm wrappers.

Wrapper types with bounds checks have been attempted to move that
checking to setters as well instead of residing entirely in the wrapper
types but that decision might be reversed.