.cbor spec support
Supports `bytes .cbor T` which allows encoding bytes that must obey the
format of a serialized `T`. This is supported both in fields and
definitions. In the generated API any `bytes .cbor T` is treated as `T`
but in the serialization code it is treated as bytes.

Also removes the need to generate wrapper types for tagged types and
instead only uses aliases as they are essentially only a serialization
detail.

Includes a fix for the unit tests core/core_no_wasm as there was a low
chance of them failing as the tests seem to be run asynchronously which
could allow one of them to rarely have unit tests copied twice.

This replaces the PR #47 to avoid a bunch of git merging issues due to
it being based partway through #37 before it was merged and in a fork.