JSON Support
--json-serde-derives

Allows to/from JSON support for all types. In Rust this works via the
serde traits. In Wasm there are explicit to_json/to_json_value/from_json
functions added directly to the wasm wrapper's impl

--json-schema-export

Derives schemars::JsonSchema for all types, which allows exporting of
the serde JSON schema to an external directory. This generates a 3rd
crate, json-gen, which when run exports them in .json format to
json-gen/schemas/

When preserve-encodings is enabled, to facilitate the above
auto-derives, we are replacing linked_hash_map::LinkedHashMap with a
wrapper type exported to src/ordered_hash_map.rs to get around the rust
orphan rule.