package.json generation
`--package-json`

Generates a NPM package along with build scripts to build the project.
When using this option all crates will be in a /rust/ directory instead
of the project root.

If `--json-schema-export` is enabled then it will also include scripts
for including those JSON schemas into the generated .ts file.

Also includes a fix for to/from bytes methods in wasm builds when
enabled