@no_alias tag + to/from bytes rework + misc fixes + external types (#130)
* @no_alias tag + to/from bytes rework + misc fixes

Aliases are not generated if you tag comments by ; @no_alias
This is useful for example in babbage.cddl's `int64 =
-9223372036854775808 .. 9223372036854775807 ; @no_alias` to avoid
generating this.

The to/from bytes were reworked to have both a `to_cbor_bytes()` and
`to_canonical_cbor_bytes()` instead a possibly more confusing bool
param. `from_bytes()` is renamed to `from_cbor_bytes()` as well to make
it more explicit. This helps with how CML is implemented right now to
match that code better.

ToBytes trait renamed to ToCBORBytes and is only used for
non-canonical-preserve builds so we can avoid using
cbor_event::Serialize as well as our own Serialize when not needed.

Misc fixes:
* bool serialization changed to not rely on cbor_event::Serialize
* tests updated for all changes + removed old TODO
* AsRef trait now generated for wasm wrappers to help for by-ref access
  to the underlying struct without needing to clone it with Into

* AsRef for table/array WASM types too

This also removes the pub (crate) visibility as it's no longer needed
and not really desired anyway as we're in a multi-crate structure in CML
right now and AsRef is basically the same functionality-wise.

* Tests for external defs

Tests for external defs in the core test (wasm + non-wasm builds) that
pulls in externally coded files (renamed codegen results of the `foo`
definition as `ExternalFoo`).

Error when combining `@newtype` and `@no_alias`

Removed a couple unused lines of code