_CDDL_CODEGEN_RAW_BYTES_TYPE_ support (#148)
This is used for generating CML mostly as it lets us use existing
structs, similar to `_CDDL_CODEGEN_EXTERN_`, but with the difference being
that instead of simply calling the `Serialize`/`Deserialize` trait it
instead treats it as if it's bytes by converting using CML's
`RawBytesEncoding` trait. This is incredibly useful for crypto-related
types (keys, hashes, etc) that don't store encoding details. Here, the
encoding details are stored where they're used as if they were using
`bytes` directly.