Fix issue with overlapping inlined basic embedded enums
Caused when the inlining would made cddl-codegen think that the types
were not overlapping since it was looking at the stored type not the
actual starting cbor type (e.g. when it was a fixed value).

This would cause a problem as the type matching introduced in #199
but only in very specific cases with basic groups starting with fixed
values.